Added IAnimal and seperate files for classes
